WebAPI/requirements.txt
Flask==3.0.1
Jinja2==3.1.3
pdfkit==1.0.0
pydantic==2.6.0
python-dotenv==1.0.1
pipreqs==0.4.13